Refactoring code from if else to ternary operators
I used ternary operators to check each condition and perform the corresponding actions. For the second condition, in which the player hits the secret number, we have several actions to be performed, so I used the operators , to separate them into a single expression. In case score > highscore the maximum score is updated otherwise we don't need to do anything so we use null. For the third condition, when the guess is incorrect, we also have two different actions depending on the score value, so again we use the ternary operator to decide which action to take.